+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -38,13 +38,7 @@
                 type="Magento\InventoryCatalog\Plugin\CatalogInventory\UpdateSourceItemAtLegacyStockItemSavePlugin"/>
     </type>
     <type name="Magento\CatalogInventory\Model\ResourceModel\Stock\Status">
-        <plugin name="adapt_add_stock_data_to_collection_to_multi_stocks"
-                type="Magento\InventoryCatalog\Plugin\CatalogInventory\AdaptAddStockDataToCollectionToMultiStocks"/>
-    </type>
-    <preference for="Magento\InventoryCatalog\Model\GetStockIdForCurrentWebsiteInterface"
-                type="Magento\InventoryCatalog\Model\GetStockIdForCurrentWebsite"/>
-    <type name="Magento\CatalogInventory\Model\ResourceModel\Stock\Status">
-        <plugin name="adapt_add_stock_status_to_select_to_multi_stocks"
-                type="Magento\InventoryCatalog\Plugin\CatalogInventory\AdaptAddStockStatusToSelectToMultiStocks"/>
+        <plugin name="adapt_resource_model_stock_status_to_multi_stocks"
+                type="Magento\InventoryCatalog\Plugin\CatalogInventory\AdaptResourceModelStockStatusToMultiStocks"/>
     </type>
 </config>
